refactor: sync 패키지 timer 패키지 안으로 이동
