--- backend/src/main/java/site/coduo/timer/exception/NegativeTimeException.java
@@ -1,8 +1,0 @@
-package site.coduo.timer.exception;
-
-public class NegativeTimeException extends TimerException{
-
-    public NegativeTimeException(final String message) {
-        super(message);
-    }
-}
