+++ backend/src/main/java/site/coduo/timer/exception/NotFoundTimeStampException.java
@@ -1,4 +1,4 @@
-package site.coduo.sync.exception;
+package site.coduo.timer.exception;
 
 public class NotFoundTimeStampException extends SyncException {
