+++ backend/src/main/java/site/coduo/timer/exception/SyncException.java
@@ -1,4 +1,4 @@
-package site.coduo.sync.exception;
+package site.coduo.timer.exception;
 
 public class SyncException extends RuntimeException {
