+++ backend/src/main/java/site/coduo/timer/service/TimestampRegistry.java
@@ -6,8 +6,8 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.springframework.stereotype.Component;
 
 import lombok.NoArgsConstructor;
-import site.coduo.sync.exception.NotFoundTimeStampException;
 import site.coduo.timer.domain.Timer;
+import site.coduo.timer.exception.NotFoundTimeStampException;
 
 @Component
 @NoArgsConstructor
