+++ backend/src/main/java/site/coduo/websocket/stomp/StompEventListener.java
@@ -8,7 +8,7 @@ import org.springframework.web.socket.messaging.SessionSubscribeEvent;
 import org.springframework.web.socket.messaging.SessionUnsubscribeEvent;
 
 import lombok.RequiredArgsConstructor;
-import site.coduo.sync.service.SchedulerService;
+import site.coduo.timer.service.SchedulerService;
 import site.coduo.websocket.exception.NotFoundAccessCodeInQueryException;
 
 @Component
@@ -16,6 +16,7 @@ import site.coduo.websocket.exception.NotFoundAccessCodeInQueryException;
 public class StompEventListener {
 
     private static final int DESTINATION_PREFIX_LENGTH = "/topic/".length();
+    private static final String PATH = "/";
 
     private final SchedulerService schedulerService;
 
@@ -24,7 +25,7 @@ public class StompEventListener {
         final StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
         final String destination = (String) headerAccessor.getHeader(SimpMessageHeaderAccessor.DESTINATION_HEADER);
         if (destination == null) {
-            throw new NotFoundAccessCodeInQueryException("STOMP 헤더에 simpDestination가 존재하지 않습니다.");
+            throw new NotFoundAccessCodeInQueryException("STOMP 헤더에 simpDestination이 존재하지 않습니다.");
         }
         final String key = parsePairRoomKey(destination);
         schedulerService.notifyTimerStatus(key);
@@ -42,7 +43,7 @@ public class StompEventListener {
     }
 
     private String parsePairRoomKey(final String destination) {
-        final int endIndex = destination.indexOf("/", DESTINATION_PREFIX_LENGTH);
+        final int endIndex = destination.indexOf(PATH, DESTINATION_PREFIX_LENGTH);
         if (endIndex == -1) {
             throw new NotFoundAccessCodeInQueryException("STOMP subscribe의 destination에서 accessCode를 파싱할 수 없습니다.");
         }
